Refactored architecture to be more scalable. Renamed classes to be more suitable
