+++ client/Assets/Entities.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: ba1d43d3a5bc0a04cbc71c58b2107e23
+folderAsset: yes
+timeCreated: 1467058382
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
