+++ client/Assets/Entities/Base.cs.meta
