+++ client/Assets/Entities/Players.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: add3db5545db6e3499d45fa8fc0e2dd4
+folderAsset: yes
+timeCreated: 1467058382
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
