+++ client/Assets/Entities/Players/ConnectedPlayer.cs.meta
